requirements.txt
django-emoji
asgiref==3.2.3
bcrypt==3.1.7
beautifulsoup4==4.8.2
cffi==1.13.2
cheroot==8.2.1
dj-database-url==0.5.0
Django==2.2.6
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-froala-editor==3.0.6
django-heroku==0.3.1
dnspython==1.16.0
gunicorn==20.0.4
jaraco.functools==2.0
Markdown==3.1.1
more-itertools==8.1.0
psycopg2==2.8.4
pycparser==2.19
pymongo==3.10.1
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
web.py==0.40
whitenoise==5.0.1
